fix: move fns to custom hook and utils. off linter issue
